Combined AdType.java with DemoMenuItem.java
