--- AppLovin MAX Demo App - Java/app/src/main/java/com/applovin/enterprise/apps/demoapp/ads/applovin/DemoMenuItem.java
@@ -1,25 +1,0 @@
-package com.applovin.enterprise.apps.demoapp.ads.applovin;
-
-import android.content.Intent;
-
-public class DemoMenuItem
-{
-    private final String title;
-    private final Intent intent;
-
-    public DemoMenuItem(final String title, final Intent intent)
-    {
-        this.title = title;
-        this.intent = intent;
-    }
-
-    public String getTitle()
-    {
-        return title;
-    }
-
-    public Intent getIntent()
-    {
-        return intent;
-    }
-}
