+++ AppLovin MAX Demo App - Java/app/src/main/java/com/applovin/enterprise/apps/demoapp/ads/applovin/interstitials/InterstitialDemoMenuActivity.java
@@ -3,7 +3,7 @@ package com.applovin.enterprise.apps.demoapp.ads.applovin.interstitials;
 import android.content.Intent;
 
 import com.applovin.enterprise.apps.demoapp.ads.applovin.DemoMenuActivity;
-import com.applovin.enterprise.apps.demoapp.ads.applovin.DemoMenuItem;
+import com.applovin.enterprise.apps.demoapp.data.main.DemoMenuItem;
 
 public class InterstitialDemoMenuActivity
         extends DemoMenuActivity
