+++ AppLovin MAX Demo App - Java/app/src/main/java/com/applovin/enterprise/apps/demoapp/ads/applovin/mrecs/MRecDemoMenuActivity.java
@@ -3,7 +3,7 @@ package com.applovin.enterprise.apps.demoapp.ads.applovin.mrecs;
 import android.content.Intent;
 
 import com.applovin.enterprise.apps.demoapp.ads.applovin.DemoMenuActivity;
-import com.applovin.enterprise.apps.demoapp.ads.applovin.DemoMenuItem;
+import com.applovin.enterprise.apps.demoapp.data.main.DemoMenuItem;
 
 /**
  * Created by monica on 7/24/17.
